requirements.txt
# AWS SDK for Bedrock
boto3>=1.34.0
python-dotenv>=1.0.0

# PDF Generation
reportlab>=4.0.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
mangum>=0.17.0  # For Vercel serverless deployment

# Audio Processing
pydub>=0.25.1

# Testing
pytest>=7.4.0
hypothesis>=6.92.0
